chore: rename demo files
