+++ packages/nutui-taro-demo/src/dentry/pages/cascader/index.vue
@@ -3,42 +3,42 @@
     <h2>{{ t('basic') }}</h2>
     <Basic />
 
-    <h2>{{ t('customAttributeName') }}</h2>
-    <CustomAttributeName />
+    <h2>{{ t('custom') }}</h2>
+    <Custom />
 
-    <h2>{{ t('asyncLoading') }}</h2>
-    <AsyncLoading />
+    <h2>{{ t('lazy') }}</h2>
+    <Lazy />
 
-    <h2>{{ t('asyncLoadingPartial') }}</h2>
-    <AsyncLoadingPartial />
+    <h2>{{ t('partial') }}</h2>
+    <Partial />
 
-    <h2>{{ t('automaticTransfer') }}</h2>
-    <AutomaticTransfer />
+    <h2>{{ t('convert') }}</h2>
+    <Convert />
   </Demo>
 </template>
 
 <script setup lang="ts">
 import { useTranslate } from '../../../utils';
 import Basic from './basic.vue';
-import CustomAttributeName from './custom-attribute-name.vue';
-import AsyncLoading from './async-loading.vue';
-import AsyncLoadingPartial from './async-loading-partial.vue';
-import AutomaticTransfer from './automatic-transfer.vue';
+import Custom from './custom.vue';
+import Lazy from './lazy.vue';
+import Partial from './partial.vue';
+import Convert from './convert.vue';
 
 const t = useTranslate({
   'zh-CN': {
     basic: '基础用法',
-    customAttributeName: '自定义属性名称',
-    asyncLoading: '动态加载',
-    asyncLoadingPartial: '部分数据动态加载',
-    automaticTransfer: '自动转换'
+    custom: '自定义属性名称',
+    lazy: '动态加载',
+    partial: '部分数据动态加载',
+    convert: '自动转换'
   },
   'en-US': {
     basic: 'Basic Usage',
-    customAttributeName: 'Custom attribute name',
-    asyncLoading: 'Async loading',
-    asyncLoadingPartial: 'Async loading of partial data',
-    automaticTransfer: 'Automatic data conversion'
+    custom: 'Custom attribute name',
+    lazy: 'Async loading',
+    partial: 'Async loading of partial data',
+    convert: 'Automatic data conversion'
   }
 });
 </script>
